Attempt to fix map scrolling issue with uint16 for map tile itx -e Original/build commit ID: 9c5cf516b0a2b9439741c9b3ca3589ab624670d6

+++ main.c
@@ -218,9 +218,9 @@ void move_background(signed int move_x, signed int move_y)
     unsigned int itx_y;
     unsigned int base_itx_x;
     unsigned int base_itx_y;
-    unsigned int current_tile_itx;
-    unsigned int current_tile_data_itx;
-    unsigned int current_tile_palette_itx;
+    UINT16 current_tile_itx;
+    UINT16 current_tile_data_itx;
+    UINT16 current_tile_palette_itx;
     unsigned int itx_x_max;
     unsigned int itx_y_max;
     signed int direction_tile_offset_x;
